Refactor code and enhanced readability of tests
The main library app code was significantly refactored for creating books and this includes the addition of a new function, createBook. Also, `DisplayName` annotations with meaningful details were added to make the test methods in the BookTest more understandable and correct a typo in LibraryTest. An assertion was also fixed in FictionTest.